Added social media information
Based on the information from https://github.com/InnerSourceCommons/innersourcecommons.org/issues/110
I added social media information.

I also translated the GitHub title from German to English and
removed the commented ones (we can re-add them later when they are
we have actually have them)

I am not 100% happy with having two LinkedIn icons, but one is for
the page and another one is for the group, but this can be rediscussed.