Added event - InnerSource day at OSCON 2019
I added InnerSource day at OSCON 2019 to past events.
I also did a minor adjustment to the text of ISC Summit 5, adding
an extra dash to have the same patters as the others.

I took the event url and created a snapshot at web.archive.org, to
prevent from having a broken link in the future (it may break, but we have
a fallback :D)
This practice should become a Best-practice (maybe add it to the contribution
guideline).