Fix navigation bug.
The previous approach (using jQuery) resulted in the top bar navigation failing to show the dropdown menus on focus. Loading jQuery in the header rather than footer was the root cause (presumably there is some hijacking of the focus event). Reverted change to where jQuery is loaded and used  instead (drawback is no compatibility with Internet Explorer).