Self host articles (#159)
* Enable article embedding.

- Bring in asciidoctor.min.js
- Move javascript.min.js to header
- Embed article retrieved from raw GitHub.

* Embed Contributor articles.

* Fix navigation bug.

The previous approach (using jQuery) resulted in the top bar navigation failing to show the dropdown menus on focus. Loading jQuery in the header rather than footer was the root cause (presumably there is some hijacking of the focus event). Reverted change to where jQuery is loaded and used  instead (drawback is no compatibility with Internet Explorer).

* Use asciidoctor.js from CDN.

* Use new layout for learning path pages rather than overloading page.

* Remove debugger statements.

* Pull in asciidoctor.js from learning path page template instead of conditional in head.

* Generate learning path menus from front matter.

* Add previous and next links to bottom of learning path pages.

* Add remaining sections of the learning path.

* Add back link.