Update isc-fall-2020-speakers.md
Added in additional speakers. 
there seems to be a bug in that an "i" appears before Tim's, Sebastian's and Tom's names. But I can't seem to get rid of it. 
Also, be nicer if the names appear under the pics (unlike how it is appearing in the preview) - not sure if that will resolve itself....